+++ CHANGELOG.md
@@ -41,7 +41,7 @@ All user visible changes to this project will be documented in this file. This p
     - `PeerConnection` renegotiation functionality ([#105]);
     - Calculate and send call quality score based on RTC stats ([#132]);
     - Muting/unmuting `MediaTrack`s by receiver ([#127]);
-    - Send `TrackUpdate::IceRestart` on `PeerConnection.connectionState` `Failed` ([#138]).
+    - Send `TrackUpdate::IceRestart` on based on RTC stats analysis ([#138]).
 - [Coturn] integration:
     - [Coturn] sessions destroying ([#84]);
     - [Coturn] stats processing ([#94]).